# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: TAIS - Terminal AI System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plugins:
    get:
      operationId: listPlugins
      tags: [plugins]
      summary: List all plugins
      responses:
        "200":
          description: List of plugins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plugin"
    post:
      operationId: installPlugin
      tags: [plugins]
      summary: Install a plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginInput"
      responses:
        "201":
          description: Installed plugin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
  /plugins/{id}:
    get:
      operationId: getPlugin
      tags: [plugins]
      summary: Get plugin by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plugin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlugin
      tags: [plugins]
      summary: Delete a plugin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /plugins/{id}/toggle:
    patch:
      operationId: togglePlugin
      tags: [plugins]
      summary: Toggle plugin enabled/disabled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated plugin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List all registered models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiModel"
    post:
      operationId: registerModel
      tags: [models]
      summary: Register a new AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiModelInput"
      responses:
        "201":
          description: Registered model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
  /models/active:
    get:
      operationId: getActiveModel
      tags: [models]
      summary: Get the currently active model
      responses:
        "200":
          description: Active model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
        "404":
          description: No active model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /models/{id}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get model by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete a model record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /models/{id}/activate:
    patch:
      operationId: activateModel
      tags: [models]
      summary: Set model as active
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PluginManifest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        author:
          type: string
        permissions:
          type: array
          items:
            type: string
        triggers:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: string
        minAppVersion:
          type: string
        homepage:
          type: ["string", "null"]
      required:
        - id
        - name
        - version
        - description
        - permissions
        - triggers
        - actions
    Plugin:
      type: object
      properties:
        id:
          type: integer
        pluginId:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        author:
          type: ["string", "null"]
        enabled:
          type: boolean
        manifest:
          $ref: "#/components/schemas/PluginManifest"
        installedAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - pluginId
        - name
        - version
        - description
        - enabled
        - manifest
        - installedAt
        - updatedAt
    PluginInput:
      type: object
      properties:
        manifest:
          $ref: "#/components/schemas/PluginManifest"
        enabled:
          type: boolean
      required:
        - manifest
    AiModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        provider:
          type: string
          enum: [qwen, gemma, phi, smollm, tinyllama, custom]
        filename:
          type: string
        localPath:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        sizeBytes:
          type: ["integer", "null"]
        quantization:
          type: ["string", "null"]
        contextLength:
          type: ["integer", "null"]
        isActive:
          type: boolean
        status:
          type: string
          enum: [available, downloading, error, missing]
        downloadProgress:
          type: ["integer", "null"]
        installedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - provider
        - filename
        - isActive
        - status
        - createdAt
    AiModelInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        provider:
          type: string
          enum: [qwen, gemma, phi, smollm, tinyllama, custom]
        filename:
          type: string
          minLength: 1
        localPath:
          type: string
        downloadUrl:
          type: string
        sizeBytes:
          type: integer
        quantization:
          type: string
        contextLength:
          type: integer
      required:
        - name
        - provider
        - filename
...